+++ com.unity.render-pipelines.high-definition/HDRP/Lighting/LightLoop/LightLoop.cs
@@ -1443,9 +1443,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         {
             m_ShadowMgr.UpdateCullingParameters(ref cullingParams);
             // In HDRP we don't need per object light/probe info so we disable the native code that handles it.
-            #if UNITY_2018_2_OR_NEWER
             cullingParams.cullingFlags |= CullFlag.DisablePerObjectCulling;
-            #endif
         }
 
         public bool IsBakedShadowMaskLight(Light light)